feat(kilo-chat-hooks): extract useMessageCacheUpdater into shared package
Moves the live event-stream cache patcher from the web-only useMessages
file into @kilocode/kilo-chat-hooks. Adds an optional onActionFailed
callback so platform wrappers inject toasts; web passes toast.error.
